refactor: Better data structure
