+++ packages/cozy-scanner/src/DocumentQualification.jsx
@@ -10,7 +10,9 @@ import Grid from 'cozy-ui/transpiled/react/MuiCozyTheme/Grid'
 import CategoryGridItem from './CategoryGridItem'
 import DocumentCategory from './DocumentCategory'
 
-import { themes, getItemById, getItemsByCategory } from './DocumentTypeData'
+import { themes } from './DocumentTypeData'
+import { getItemById } from './DocumentTypeDataHelpers'
+import { getItemsByCategory } from './DocumentTypeDataHelpers'
 import GridItem from 'cozy-ui/transpiled/react/Labs/GridItem'
 import styles from './styles.styl'
